Turn graph node into class and update dependents
By turning the graph node into a class, it's easier to define what makes
a graph node unique (we just create a key function that uniquely
identifies it). Search is updated accordingly to accomodate this change.
Simulation is minorly updated to process the return of the BFS and
returns a dummy ExecutionTrace[].