Create trace for FSA return and fix bugs (usable!)
This commit is the first complete iteration of the new graph search
algorithm.

A 'read' property on graph nodes was added to store what symbol was
read during a transition, and the 'read' property of type FSAGraphState
was changed to remaining to better reflect its purpose.

A bug was fixed on line 130->142 of graph.ts which erroneously threw
away a potentially valid transition.