Rework simulation algorithm to be even more generic
In the previous implementation, there was no enforced relationship
between the states, transitions and nodes of a graph being solved. The
changes in this commit aim to rectify that with the generic type
parameter enforcements. This reduces code redundancy and guarantees
sound execution of the graph search.

Furthermore, the existence of a null check on the initial state meant
that graphs could be initialised in an error state. Graphs must now have
at least an initial state before the search commences.