Implement generic stepper, change State to class
Changing the state to a class meant that I could enforce the existence
of a key function without it having to exist already on the front end.
The stepper algorithm currently steps too far forward, but it should be
good for testing.

There's a bit of redundant code in simulateFSA.ts we would have to
tackle. This will likely involve making some assurances on the front end
about the shape of the data passed to this package.