Move to using `resolveGraph` instead of other versions
I had to make the typing weaker to deal with everything, but will improve it in a later refactor. The main problem is that Turing machines have a single read character instead of multiple like other machines. But only reason I need to deal with this anyways is because the frontend doesn't use proper classes