1 morbillion increase in performance
`useActions` was getting called hundred of times when ever there was a rerender (Should only be called once on page load). This caused massive slowdowns and would mean that useActions wouldn't scale.

Most of the problems have been fixed by early returning in hidden components which were calling `useActions`. The problem still isn't fully fixed because the Editor calls it on every rerender but that doesn't happen as much anymore so shouldn't be too much of a problem