Forget what has been selected if user is selecting a new item
The fix works by checking if the user is just "clicking" (this is defined as just left clicking without holding the shift key). If they are we then see if they are selecting something that isn't already selected, and clearing the memory of selected items if so.

Holding shift will stay add the item to the selection.