Fix TM Tape Stretching Components (#441)
Closes #440.

Issue was caused by the whole tape taking space at all times and the
animation being handled by a CSS transition (putting `margin-left:
1000px` on an element will get a similar effect).

This is a bandaid fix on the problem so that the TM tape for longer
input strings is not obnoxious. A rewrite/redesign of the TM tape
component might be needed to fix it completely. The fix will cull the
size of the rendered tape so that it won't stretch the middle
components, and maintain the animation (but janky) to keep up the
appearance of the tape moving.

Alternative would be to make all the tape's parents not stretch but I
couldn't for the life of me get it working (and you will still have the
overflow problem).