Template thumbnails (#442)
Closes #415. Renders the selected elements when adding a template into
the thumbnail store.

Currently it is creating a copy of the elements that are selected into a
new SVG element, but it will flash on the screen for a brief moment. Not
sure how to fix that without messing up the containers.

Other things that were changed:
- Template thumbnails use the same store as project thumbnails with a
`tmp` added at the start
- Modified the project card (which the templates use) to have a bin icon
for deleting a template. This action should also delete delete the
thumbnail from store (otherwise a garbage cleanup-esque routine like
done with the project thumbnails causes a stack overflow)
- The template ghost appears while a template is selected and follows
the mouse

Other things to be fixed:
- The template svg element flashing on the screen
- Comments not being rendered as `useImageExport` does not include
comments