Improve transition edge UX (#458)
Closes #451, closes #450.

Implements fixes in the above issues. Lumped these together as they both
modify the edge.

- Transition dragging circles have a 10 radius (20 diametre) invisible
circle as clicking error, the same thickness as the edge's invisible
path. This is drawn on top of all the edge components so take priority.
- Clicking on an edge now always selects all transitions on the edge.
- This gives a "select all transitions on edge" option so they can
select all then deselect vs selecting each individually
- A user right-clicking an edge can be assumed they want to act on that
edge, but we can add a check for right-click if that assumption is
incorrect
- Added extra context items to the edge context menu
  - Flip edge
- Edit first transition (Double-clicking the edge already edits the
first transition)
- Changed transition actions to use the context passing system so that
the targeted item is the one being acted on
- Users pressing the `delete` key while in the context menu runs the
global delete action, this keeps the menu open so added an event to
close it