Stop transition editing deleting transition on cancel (#459)
Closes #449.

Edit transition receives an optional flag specifying whether a transtion
being edited is a new transition or not. The input dialog receives that
and does not go through with he deletion if the edit event is not for a
new dialog. Defaults to `true`, which is the previous behaviour, so
should not regress on any other part of the app.