Deployment of Group 6's Work – "Modules" Feature (#508)
Hey!

We are Group 6, and we've implemented a new "**Modules**" feature for
Automatarium. This feature is designed as a substitute for traditional
tutorial sheets and is intended to provide a structured, interactive
experience for students.

A full breakdown of the feature's functionality, design, and
implementation can be found in our technical report
[here.](https://github.com/user-attachments/files/17690736/Group_6_Technical_Report.pdf)

We are happy to answer any questions or address any feedback you may
have.

Thank you for the opportunity to contribute to Automatarium!

---------

Co-authored-by: s3944192 <s3944192@student.rmit.edu.au>
Co-authored-by: ZiChengHongRPO <s3882476@student.edu.au>
Co-authored-by: Shourya <s3929838@student.rmit.edu.au>
Co-authored-by: s3948240 <s3948240@student.rmit.edu.au>
Co-authored-by: Brian Tran <minhtritran.dev@gmail.com>
Co-authored-by: Liam Keenan <s3943923@student.rmit.edu.au>
Co-authored-by: s3944192 <103025594+s3944192@users.noreply.github.com>
Co-authored-by: Dib <79500311+tdib@users.noreply.github.com>
Co-authored-by: Jake Leahy <jake@leahy.dev>